requirements.txt
python-telegram-bot[job-queue]==22.7
SQLAlchemy==2.0.36
PyMySQL==1.1.1
cryptography==42.0.5
pyrogram==2.0.106
TgCrypto==1.2.5
backoff==2.2.1
python-dotenv==1.0.0
